user-managment-REST/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.0.7
djangorestframework==3.11.0
idna==3.4
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
